rename action file
