+++ workspaces/azure-devops/plugins/scaffolder-backend-module-azure-pipeline-actions/src/module.ts
@@ -18,7 +18,7 @@ import {
   createBackendModule,
 } from '@backstage/backend-plugin-api';
 import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node/alpha';
-import { runAzurePipelineAction } from './actions/example/example';
+import { runAzurePipelineAction } from './actions/example/run-pipeline';
 import { ScmIntegrations } from '@backstage/integration';
 
 /**
@@ -35,9 +35,7 @@ export const scaffolderModule = createBackendModule({
       },
       async init({ scaffolderActions, config }) {
         const integrations = ScmIntegrations.fromConfig(config);
-        scaffolderActions.addActions(
-          runAzurePipelineAction({ integrations, config }),
-        );
+        scaffolderActions.addActions(runAzurePipelineAction({ integrations }));
       },
     });
   },
